- Moved annotator object code to trait - Created base importers - Created easy methods - Cloud models for further annotators
