+++ src/main/scala/com/johnsnowlabs/downloader/en/models/Models.scala
@@ -5,6 +5,8 @@ import com.johnsnowlabs.nlp.annotators
 import com.johnsnowlabs.nlp.annotators.ner.crf.NerCrfModel
 import com.johnsnowlabs.nlp.annotators.pos.perceptron.PerceptronModel
 import com.johnsnowlabs.downloader.ResourceDownloader
+import com.johnsnowlabs.nlp.annotator.{NorvigSweetingModel, ViveknSentimentModel}
+import com.johnsnowlabs.nlp.annotators.Lemmatizer
 
 object CloudDocumentAssembler {
   def retrieveStandard = ResourceDownloader.downloadModel(nlp.DocumentAssembler, "document_std", Some("en"))
@@ -24,5 +26,17 @@ object CloudPerceptronModel {
 
 object CloudNerCrfModel {
   def retrieveSmall = ResourceDownloader.downloadModel(NerCrfModel, "ner_fast", Some("en"))
+}
+
+object CloudLemmatizer {
+  def retrieveSmall = ResourceDownloader.downloadModel(Lemmatizer, "lemma_fast", Some("en"))
+}
+
+object CloudSpellChecker {
+  def retrieveSmall = ResourceDownloader.downloadModel(NorvigSweetingModel, "spell_fast", Some("en"))
+}
+
+object CloudViveknSentiment {
+  def retrieveSmall = ResourceDownloader.downloadModel(ViveknSentimentModel, "vivekn_fast", Some("en"))
 }
