+++ src/main/scala/com/johnsnowlabs/nlp/annotators/spell/norvig/NorvigSweetingApproach.scala
@@ -75,6 +75,8 @@ class NorvigSweetingApproach(override val uid: String)
         ResourceHelper.wordCount($(corpus), p = recursivePipeline).toMap
       } else {
         import ResourceHelper.spark.implicits._
+        dataset.show()
+        dataset.select($(inputCols).head).show
         dataset.select($(inputCols).head).as[Array[Annotation]]
           .flatMap(_.map(_.result))
           .groupBy("value").count
